+++ packages/item/src/util/theme.js
@@ -19,6 +19,7 @@ export const themeNamespace = '@atlaskit-shared-theme/item';
 // Used for Group titles and Item descriptions
 export const smallFontSize = 12;
 export const compactSmallFontSize = 10;
+export const gridSize: number = akGridSizeUnitless;
 
 export const defaultTheme: ItemTheme = {
   borderRadius: parseInt(akBorderRadius, 10),