+++ packages/task-decision/src/support/MockTaskDecisionResource.ts
@@ -18,6 +18,8 @@ import {
   ItemResponse,
   ObjectKey,
   Query,
+  RecentUpdateContext,
+  RecentUpdatesId,
   TaskDecisionProvider,
   TaskResponse,
   TaskState,
@@ -63,6 +65,14 @@ export default class MockTaskDecisionResource implements TaskDecisionProvider {
     const serviceItemResponse = getServiceItemsResponse();
     const result = convertServiceItemResponseToItemResponse(serviceItemResponse);
     return this.applyConfig(query, result, 'items');
+  }
+
+  unsubscribeRecentUpdates(id: RecentUpdatesId) {
+
+  }
+
+  notifyRecentUpdates(updateContext?: RecentUpdateContext) {
+
   }
 
   private getNextDate() {