+++ packages/task-decision/src/type-helpers.ts
@@ -2,6 +2,7 @@ import {
   Decision,
   Item,
   ObjectKey,
+  RendererContext,
   ServiceDecision,
   ServiceItem,
   ServiceTask,
@@ -17,11 +18,19 @@ export const isServiceTask = (item: ServiceItem): item is ServiceTask => !!(item
 
 export const isDateSortCriteria = (sortCriteria?: SortCriteria) => !sortCriteria || sortCriteria === 'creationDate' || sortCriteria === 'lastUpdateDate';
 
-export const toObjectKey = (item: Item): ObjectKey => {
+export const toObjectKey = (item: Item | ServiceDecision | ServiceTask): ObjectKey => {
   const { containerAri, localId, objectAri } = item;
   return {
     containerAri,
     localId,
+    objectAri
+  };
+};
+
+export const toRendererContext = (item: Item | ObjectKey): RendererContext => {
+  const { containerAri, objectAri} = item;
+  return {
+    containerAri,
     objectAri
   };
 };